GeneralActionModule accepts all kind of lambda now.
Tested and approuved !

YYYYEEEESSSS !! I succeeded in making an object that accepts all lambdas
and such ! I'm happy ! It's a really huge step in my programming
competences ! Now, I can take as many lambda as I want with a lot of
parameter in the capture and pass it into my class that encapsulate this
lambda... I'm happyyyy, reeeaaallllyyyy !!! ^^

Now, a true general action class was born here; Functional, simple,
adaptative, optimized. Really beautiful !
Now, I won't stay at the simple level of a function pointer that forbid
the capture of a lambda and limits the possibilities of a changes in the
parameters. Hahahahahaha ^^ !!!

Sayonara, function pointer ! You serve me right but now, I must part with
you... forever ^^.